backend/requirements.txt
fastapi
uvicorn
networkx
pydantic
python-dotenv
requests